requirements.txt
numpy>=1.13.0,<2
Pillow>=4.1.1,<5
joblib>=0.11,<1
scikit-image>=0.13.0,<1
